requirements.txt
numpy
Flask
gunicorn
pandas
PyYAML
evidently
dill
numpy
matplotlib
requests==2.28.1
jupyterlab==3.4.7  
pyspark==3.2.1
PyYAML
ipykernel==6.15.0  
-e .